Separate a few more test classes
